pyproject.toml
[build-system]
requires = ["setuptools >= 62.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebattle_october_2022_api"
version = "0.0.1"

[project.optional-dependencies]

test = [
  "pytest",
  "pytest-xdist",
  "pytest-cov",
  "pytest-django",
  "faker",
  "factory-boy",
  "tox",
  "mypy",
  "flake8",
  "pytest-factoryboy",
  "pytest-instafail",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["*tests*", "htmlcov"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79
default_section = "THIRDPARTY"
known_first_party = ["api", "users", "config"]
known_third_party = ["django", "pytest", "rest_framework", "setuptools"]

[tool.black]
line-length = 79
include = '\.pyi?%'

[tool.mypy]
mypy_path = "."
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
strict_optional = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = [
  "-n",
  "auto",
  "--cov",
  "--cov-config=./pyproject.toml",
  "--cov-branch",
  "--cov-report=html",
  "--create-db",
  "--no-migrations",
  "--pdbcls=pdbr:RichPdb",
]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
log_level = "ERROR"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
log_cli = true
log_cli_level = "ERROR"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
branch = true
omit = ["*/tests/*", "*/migrations/*"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.mypy_django_plugin]
ignore_missing_model_attributes = true

[tool.pydocstyle]
ignore = "D100,D101,D102,D103,D104,D105,D106,D107"

[tool.commitizen]
name = "cz_conventional_commits"
version_files = ["pyproject.toml", "./api/__init__.py"]
version = "0.0.1"
tag_format = "$version"

style = [
  [
    "qmark",
    "fg:#ff9d00 bold",
  ],
  [
    "question",
    "bold",
  ],
  [
    "answer",
    "fg:#ff9d00 bold",
  ],
  [
    "pointer",
    "fg:#ff9d00 bold",
  ],
  [
    "highlighted",
    "fg:#ff9d00 bold",
  ],
  [
    "selected",
    "fg:#cc5454",
  ],
  [
    "separator",
    "fg:#cc5454",
  ],
  [
    "instruction",
    "",
  ],
  [
    "text",
    "",
  ],
  [
    "disabled",
    "fg:#858585 italic",
  ],
]
